level 1, now the issue is that 10 needs to wait some more because the server he is assigned to is still busy. I think we need another method in server. So if the manager picks a waiting customer, and passes the waiting customer and his assigned server to a method called process waits.
now the process waits method will check the assigned server's queue. if more than 1 people are in queue, will check if the head of the queue is equal to this customer. if so, then do the normal assign to served

else if more than 1 person is in queue and head isn't this waiting customer, there won't be a change of status for the customer (remains as waits) but the this timing will change to the server's next available time.

if only 1 person is in queue, carry on as normal.

event registry: only register events if there's a change of status b/w when the manager picked the customer from the queue and when the customer has been modified... todo.